requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
future==0.18.2
idna==3.3
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
mesh2sdf==0.1.0
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.1
plyfile==0.7.4
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smplx==0.1.28
tensorboardX==2.5
threadpoolctl==3.1.0
tqdm==4.63.1
trimesh==3.10.7
urllib3==1.26.9
